+++ packages/react-dnd/src/hooks/useDragSource.ts
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { DragDropMonitor, SourceType } from 'dnd-core'
+import { SourceType } from 'dnd-core'
 import registerSource from '../registerSource'
 import createSourceMonitor from '../createSourceMonitor'
 import {
@@ -7,53 +7,13 @@ import {
 	DragPreviewOptions,
 	DragSourceHookSpec,
 } from '../interfaces'
+import { useDragSourceHandler } from './useDragSourceHandler'
 import { useDragDropManager } from './useDragDropManager'
 import { useMonitorSubscription } from './useMonitorSubscription'
 import { Ref, HandlerManager, isRef } from './util'
 
-function useDragSourceHandler<DragObject>(
-	sourceSpec: DragSourceHookSpec<DragObject>,
-) {
-	const sourceSpecRef = React.useRef(sourceSpec)
-
-	React.useEffect(() => {
-		sourceSpecRef.current = sourceSpec
-	})
-
-	// Can't use createSourceFactory, as semantics are different
-	const handler = React.useMemo(
-		() => ({
-			canDrag() {
-				const { canDrag } = sourceSpecRef.current
-				return canDrag ? canDrag() : true
-			},
-			isDragging(globalMonitor: DragDropMonitor, sourceId: string) {
-				const { isDragging } = sourceSpecRef.current
-				return isDragging
-					? isDragging()
-					: sourceId === globalMonitor.getSourceId()
-			},
-			beginDrag(component: any) {
-				const { beginDrag } = sourceSpecRef.current
-				if (beginDrag) {
-					beginDrag(component)
-				}
-			},
-			endDrag(component: any) {
-				const { endDrag } = sourceSpecRef.current
-				if (endDrag) {
-					endDrag(component)
-				}
-			},
-		}),
-		[],
-	)
-
-	return handler
-}
-
 /**
- *
+ * useDragSource hook
  * @param ref The drag source element
  * @param type The drag source type
  * @param sourceSpec The drag source specification
@@ -87,13 +47,11 @@ export function useDragSource<DragObject = {}>(
 	React.useEffect(function connectDragSourceToBackend() {
 		const dragSourceNode = ref.current
 		const dragSourceOptions = sourceSpec.dragSourceOptions
-		const disconnectDragSource = backend.connectDragSource(
+		return backend.connectDragSource(
 			sourceMonitor.getHandlerId(),
 			dragSourceNode,
 			dragSourceOptions,
 		)
-
-		return disconnectDragSource
 	}, [])
 
 	React.useEffect(function connectDragPreviewToBackend() {
@@ -106,13 +64,11 @@ export function useDragSource<DragObject = {}>(
 			? (sourceSpec.dragPreview as Ref<any>).current
 			: sourceSpec.dragPreview
 		const dragPreviewOptions = sourceSpec.dragPreviewOptions
-		const disconnectDragPreview = backend.connectDragPreview(
+		return backend.connectDragPreview(
 			sourceMonitor.getHandlerId(),
 			dragPreviewNode,
 			dragPreviewOptions,
 		)
-
-		return disconnectDragPreview
 	}, [])
 
 	return sourceMonitor
