+++ packages/react-dnd/src/hooks/useDropTarget.ts
@@ -6,39 +6,7 @@ import { DropTargetMonitor, DropTargetHookSpec } from '../interfaces'
 import { useDragDropManager } from './useDragDropManager'
 import { useMonitorSubscription } from './useMonitorSubscription'
 import { Ref, HandlerManager } from './util'
-
-function useDropTargetHandler(targetSpec: DropTargetHookSpec) {
-	const targetSpecRef = React.useRef(targetSpec)
-
-	React.useEffect(function updateDropTargetSpec() {
-		targetSpecRef.current = targetSpec
-	})
-
-	// Can't use createSourceFactory, as semantics are different
-	const handler = React.useMemo(
-		() => ({
-			canDrop() {
-				const { canDrop } = targetSpecRef.current
-				return canDrop ? canDrop() : true
-			},
-			hover(component: any) {
-				const { hover } = targetSpecRef.current
-				if (hover) {
-					hover(component)
-				}
-			},
-			drop(component: any) {
-				const { drop } = targetSpecRef.current
-				if (drop) {
-					drop(component)
-				}
-			},
-		}),
-		[],
-	)
-
-	return handler
-}
+import { useDropTargetHandler } from './useDropTargetHandler'
 
 /**
  * useDropTarget Hook
@@ -69,16 +37,15 @@ export function useDropTarget(
 		targetMonitor,
 	)
 
-	React.useEffect(() => {
+	React.useEffect(function connectDropTarget() {
 		const dropTargetNode = ref.current
 		if (dropTargetNode) {
 			const dropTargetOptions = targetSpec.dropTargetOptions
-			const disconnectDropTarget = backend.connectDropTarget(
+			return backend.connectDropTarget(
 				targetMonitor.getHandlerId(),
 				dropTargetNode,
 				dropTargetOptions,
 			)
-			return disconnectDropTarget
 		}
 	}, [])
